use requirements, setup paginate

+++ remove_cluster.py
@@ -1,6 +1,7 @@
 import json
 import os
 import sys
+import itertools
 sys.path.insert(0, './vendor-boto3')
 import boto3
 from util import validate_config_input, validate_unique_cluster_name
@@ -22,33 +23,7 @@ def remove_cluster(event, context):
     if validate_unique_cluster_name(cluster_name, CLUSTER_TABLE) is not None:
 
         # Remove associated user secrets
-        list_of_secrets = SECRETS_CLIENT.list_secrets()
-        delete_secrets(list_of_secrets, cluster_name)
-
-        #SECRETS_CLIENT.delete_secret(
-        #                SecretId="cloud-infra.cloud-client-certificate-data-cloud-infra.cloud",
-        #                ForceDeleteWithoutRecovery=True
-        #            )
-        #SECRETS_CLIENT.delete_secret(
-        #                SecretId="cloud-infra.cloud-client-key-data-cloud-infra.cloud",
-        #                ForceDeleteWithoutRecovery=True
-        #            )
-        #SECRETS_CLIENT.delete_secret(
-        #                SecretId="cloud-infra.cloud-username-cloud-infra.cloud",
-        #                ForceDeleteWithoutRecovery=True
-        #            )
-        #SECRETS_CLIENT.delete_secret(
-        #                SecretId="cloud-infra.cloud-password-cloud-infra.cloud",
-        #                ForceDeleteWithoutRecovery=True
-        #            )
-        #SECRETS_CLIENT.delete_secret(
-        #                SecretId="cloud-infra.cloud-basic-auth-username-cloud-infra.cloud",
-        #                ForceDeleteWithoutRecovery=True
-        #           )
-        #SECRETS_CLIENT.delete_secret(
-        #                SecretId="cloud-infra.cloud-basic-auth-password-cloud-infra.cloud",
-        #                ForceDeleteWithoutRecovery=True
-        #            )
+        delete_secrets(cluster_name)
 
         # Remove cluster
         CLUSTER_TABLE.delete_item(
@@ -70,17 +45,27 @@ def remove_cluster(event, context):
         )
     }
 
-def delete_secrets(list_of_secrets, cluster_name):
+def delete_secrets(cluster_name):
     """Delete secrets"""
-    print(f'BOTO3 VERSION: {boto3.__version__}')
-    print(list_of_secrets)
-    for secret in list_of_secrets['SecretList']:
-        for tag in secret['Tags']:
-            if tag['Key'] == 'cluster_name' and tag['Value'] == cluster_name:
-                try:
-                    SECRETS_CLIENT.delete_secret(
-                        SecretId=f"{secret['Name']}",
-                        ForceDeleteWithoutRecovery=True
-                    )
-                except Exception as err:
-                    print(f"Secret {secret} not found, nothing to delete: {err}")
+    print("HERE")
+    for secret in paginate():
+        print(secret)
+        if 'Tags' in secret:
+            for tag in secret['Tags']:
+                if tag['Key'] == 'cluster_name' and tag['Value'] == cluster_name:
+                    try:
+                        print(f"Deleting secret: {secret['Name']}")
+                        SECRETS_CLIENT.delete_secret(
+                            SecretId=f"{secret['Name']}",
+                            ForceDeleteWithoutRecovery=True
+                        )
+                    except Exception as err:
+                        print(f"Secret {secret} not found, nothing to delete: {err}")
+
+# Doesn't work, list_secrets doest support pagninator yet :/
+def paginate():
+    client = SECRETS_CLIENT 
+    paginator = client.get_paginator("list_secrets")
+    for page in paginator.paginate().result_key_iters():
+        for result in page:
+            yield result
